requirements.txt
Flask
Flask-Login
python-dotenv
requests
